Refactored test's code to satisfy pylint's requirements
