+++ test/config_generator_test.py
@@ -1,0 +1,12 @@
+"""Test config_generator.py module"""
+from unittest import TestCase
+
+from projector_installer.config_generator import token_quote
+
+
+class ConfigGeneratorTest(TestCase):
+    """Test config_generator.py module"""
+
+    def test_token_quote(self):
+        """The token_quote method must return the same token in quotes"""
+        self.assertEqual(token_quote('some_token'), '\"some_token\"')
